Take a number's magnitude wide enough that the most negative one survives
The ceiling that sends a number too large for roman numerals or letters
back to be written in digits asked Math.Abs(number) > 32768, and
Math.Abs throws on int.MinValue because it has no positive counterpart.
So the one number furthest past that ceiling was the one number the
fallback could not catch: NumberFormatter.Format(int.MinValue, "ROMAN")
threw an OverflowException where every other out-of-range value returned
its digits.

Both guards now widen to long before taking the magnitude. Nothing
downstream changes - a number that gets past the guard is within 32768
of zero, so the Math.Abs calls that follow are safe as they stand.

Reachable because Format is public: pagination never produces a number
anywhere near this, but a caller can pass any int.

+++ MigraDocCore.DocumentObjectModel/MigraDoc.DocumentObjectModel.Fields/NumberFormatter.cs
@@ -71,7 +71,10 @@ public static class NumberFormatter
 
     static string AsRoman(int number, bool lowercase)
     {
-        if (Math.Abs(number) > 32768)
+        // Widened before the magnitude is taken, because int.MinValue has no positive counterpart
+        // and Math.Abs throws on it - so the number furthest past this ceiling was the one number
+        // the fallback could not catch.
+        if (Math.Abs((long)number) > 32768)
         {
             Debug.WriteLine(DomSR.NumberTooLargeForRoman(number), "warning");
             return number.ToString();
@@ -106,7 +109,8 @@ public static class NumberFormatter
 
     static string AsLetters(int number, bool lowercase)
     {
-        if (Math.Abs(number) > 32768)
+        // Widened for the same reason as in AsRoman above.
+        if (Math.Abs((long)number) > 32768)
         {
             Debug.WriteLine(DomSR.NumberTooLargeForLetters(number));
             return number.ToString();
